Move Labels reducer
close DCOS-14696
