+++ plugins/services/src/js/reducers/JSONMultiContainerParser.js
@@ -4,7 +4,7 @@ import {JSONParser as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONParser as environmentVariables} from './serviceForm/EnvironmentVariables';
 import {JSONParser as externalVolumes} from './serviceForm/JSONReducers/ExternalVolumes';
 import {JSONParser as healthChecks} from './serviceForm/HealthChecks';
-import {JSONParser as labels} from './serviceForm/Labels';
+import {JSONParser as labels} from './serviceForm/JSONReducers/Labels';
 import {JSONParser as localVolumes} from './serviceForm/LocalVolumes';
 import {JSONParser as portDefinitions} from './serviceForm/PortDefinitions';
 import {JSONParser as portMappings} from './serviceForm/PortMappings';
