+++ plugins/services/src/js/reducers/JSONMultiContainerReducers.js
@@ -3,7 +3,7 @@ import {JSONReducer as containers} from './serviceForm/Containers';
 import {JSONReducer as env} from './serviceForm/EnvironmentVariables';
 import {JSONReducer as fetch} from './serviceForm/JSONReducers/Artifacts';
 import {JSONReducer as scaling} from './serviceForm/MultiContainerScaling';
-import {JSONReducer as labels} from './serviceForm/Labels';
+import {JSONReducer as labels} from './serviceForm/JSONReducers/Labels';
 import {JSONReducer as volumes} from './serviceForm/MultiContainerVolumes';
 import {JSONReducer as networks} from './serviceForm/MultiContainerNetwork';
 import {JSONReducer as ipAddress} from './serviceForm/JSONReducers/IpAddress';
